Create Portfolio with 5 stock
How to analysis the market and calculate the risk in the Portfolio.

Here I took 5 stock in the single Portfoilo

i calulate the risk and up's and down's of market 

By keeping in mind there is a proverb " Don't Put All your Eggs in One Basket". This is a piece of advice which means that one should not concentrate all efforts and resources in one area as one could lose everything.   

That why i took 5 stock, if one or two goes down, 4 or 3 stock that recover the Margin.

Market is always Volatile, risk is always there but we have to find a way where risk is there but the return will be more than the expectation.

So this piece of code gives you some Knowledge.